Update single user get/update/delete paths
Also remove unused profile V1 path.
