+++ server/data/userClient.test.ts
@@ -12,33 +12,6 @@ describeClient('UserClient', provider => {
     userClient = new UserClient(token)
   })
 
-  describe('getUser', () => {
-    const user = userFactory.build()
-    const id = 'SOME_ID'
-
-    it('should return a user', async () => {
-      provider.addInteraction({
-        state: 'Server is healthy',
-        uponReceiving: 'A request to get a user',
-        withRequest: {
-          method: 'GET',
-          path: paths.users.show({ id }),
-          headers: {
-            authorization: `Bearer ${token}`,
-            'X-Service-Name': 'approved-premises',
-          },
-        },
-        willRespondWith: {
-          status: 200,
-          body: user,
-        },
-      })
-
-      const output = await userClient.getUser(id)
-      expect(output).toEqual(user)
-    })
-  })
-
   describe('getUserProfile', () => {
     const user = userFactory.build()
 
@@ -48,7 +21,7 @@ describeClient('UserClient', provider => {
         uponReceiving: 'A request to get a user profile',
         withRequest: {
           method: 'GET',
-          path: paths.users.v2profile({}),
+          path: paths.users.profile({}),
           headers: {
             authorization: `Bearer ${token}`,
             'X-Service-Name': 'approved-premises',
@@ -363,35 +336,6 @@ describeClient('UserClient', provider => {
     })
   })
 
-  describe('updateUser', () => {
-    it('should update a user', async () => {
-      const user = userFactory.build()
-
-      const rolesAndQualifications = { roles: user.roles, qualifications: user.qualifications }
-
-      provider.addInteraction({
-        state: 'Server is healthy',
-        uponReceiving: 'A request to update a user',
-        withRequest: {
-          method: 'PUT',
-          path: paths.users.update({ id: user.id }),
-          headers: {
-            authorization: `Bearer ${token}`,
-            'X-Service-Name': 'approved-premises',
-          },
-          body: rolesAndQualifications,
-        },
-        willRespondWith: {
-          status: 200,
-          body: user,
-        },
-      })
-
-      const output = await userClient.updateUser(user.id, rolesAndQualifications)
-      expect(output).toEqual(user)
-    })
-  })
-
   describe('search', () => {
     it('should search for a user', async () => {
       const users = userFactory.buildList(1)
@@ -449,30 +393,6 @@ describeClient('UserClient', provider => {
 
       const output = await userClient.searchDelius(name)
       expect(output).toEqual(user)
-    })
-  })
-
-  describe('delete', () => {
-    it('should delete a user', async () => {
-      const user = userFactory.build()
-
-      provider.addInteraction({
-        state: 'Server is healthy',
-        uponReceiving: 'A request to delete a user',
-        withRequest: {
-          method: 'DELETE',
-          path: paths.users.delete({ id: user.id }),
-          headers: {
-            authorization: `Bearer ${token}`,
-            'X-Service-Name': 'approved-premises',
-          },
-        },
-        willRespondWith: {
-          status: 200,
-        },
-      })
-
-      await userClient.delete(user.id)
     })
   })
 })
